+++ src/OpenTelemetry.Resources.AWS/AWSResourceBuilderExtensions.cs
@@ -1,6 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
+using OpenTelemetry.AWS;
 using OpenTelemetry.Internal;
 using OpenTelemetry.Resources.AWS;
 
@@ -15,10 +16,17 @@ public static class AWSResourceBuilderExtensions
     /// Enables AWS Elastic Beanstalk resource detector.
     /// </summary>
     /// <param name="builder">The <see cref="ResourceBuilder"/> being configured.</param>
+    /// <param name="configure">Optional callback action for configuring <see cref="AWSResourceBuilderOptions"/>.</param>
     /// <returns>The instance of <see cref="ResourceBuilder"/> being configured.</returns>
-    public static ResourceBuilder AddAWSEBSDetector(this ResourceBuilder builder)
+    public static ResourceBuilder AddAWSEBSDetector(this ResourceBuilder builder, Action<AWSResourceBuilderOptions>? configure = null)
     {
         Guard.ThrowIfNull(builder);
+
+        var options = new AWSResourceBuilderOptions();
+        configure?.Invoke(options);
+
+        AWSSemanticConventions.SemanticConventionVersion = options.SemanticConventionVersion;
+
         return builder.AddDetector(new AWSEBSDetector());
     }
 
@@ -26,10 +34,17 @@ public static class AWSResourceBuilderExtensions
     /// Enables AWS EC2 resource detector.
     /// </summary>
     /// <param name="builder">The <see cref="ResourceBuilder"/> being configured.</param>
+    /// <param name="configure">Optional callback action for configuring <see cref="AWSResourceBuilderOptions"/>.</param>
     /// <returns>The instance of <see cref="ResourceBuilder"/> being configured.</returns>
-    public static ResourceBuilder AddAWSEC2Detector(this ResourceBuilder builder)
+    public static ResourceBuilder AddAWSEC2Detector(this ResourceBuilder builder, Action<AWSResourceBuilderOptions>? configure = null)
     {
         Guard.ThrowIfNull(builder);
+
+        var options = new AWSResourceBuilderOptions();
+        configure?.Invoke(options);
+
+        AWSSemanticConventions.SemanticConventionVersion = options.SemanticConventionVersion;
+
         return builder.AddDetector(new AWSEC2Detector());
     }
 
@@ -38,10 +53,17 @@ public static class AWSResourceBuilderExtensions
     /// Enables AWS ECS resource detector.
     /// </summary>
     /// <param name="builder">The <see cref="ResourceBuilder"/> being configured.</param>
+    /// <param name="configure">Optional callback action for configuring <see cref="AWSResourceBuilderOptions"/>.</param>
     /// <returns>The instance of <see cref="ResourceBuilder"/> being configured.</returns>
-    public static ResourceBuilder AddAWSECSDetector(this ResourceBuilder builder)
+    public static ResourceBuilder AddAWSECSDetector(this ResourceBuilder builder, Action<AWSResourceBuilderOptions>? configure = null)
     {
         Guard.ThrowIfNull(builder);
+
+        var options = new AWSResourceBuilderOptions();
+        configure?.Invoke(options);
+
+        AWSSemanticConventions.SemanticConventionVersion = options.SemanticConventionVersion;
+
         return builder.AddDetector(new AWSECSDetector());
     }
 
@@ -49,10 +71,17 @@ public static class AWSResourceBuilderExtensions
     /// Enables AWS EKS resource detector.
     /// </summary>
     /// <param name="builder">The <see cref="ResourceBuilder"/> being configured.</param>
+    /// <param name="configure">Optional callback action for configuring <see cref="AWSResourceBuilderOptions"/>.</param>
     /// <returns>The instance of <see cref="ResourceBuilder"/> being configured.</returns>
-    public static ResourceBuilder AddAWSEKSDetector(this ResourceBuilder builder)
+    public static ResourceBuilder AddAWSEKSDetector(this ResourceBuilder builder, Action<AWSResourceBuilderOptions>? configure = null)
     {
         Guard.ThrowIfNull(builder);
+
+        var options = new AWSResourceBuilderOptions();
+        configure?.Invoke(options);
+
+        AWSSemanticConventions.SemanticConventionVersion = options.SemanticConventionVersion;
+
         return builder.AddDetector(new AWSEKSDetector());
     }
 #endif