+++ test/OpenTelemetry.Instrumentation.AWSLambda.Tests/Implementation/AWSLambdaHttpUtilsTests.cs
@@ -11,8 +11,25 @@ using Xunit;
 namespace OpenTelemetry.Instrumentation.AWSLambda.Tests.Implementation;
 
 [Collection("TracerProviderDependent")]
-public class AWSLambdaHttpUtilsTests
+public class AWSLambdaHttpUtilsTests : IDisposable
 {
+    public AWSLambdaHttpUtilsTests()
+    {
+        // have all tests use the latest Semantic Convention
+        Sdk.CreateTracerProviderBuilder()
+            .AddAWSLambdaConfigurations(opt =>
+            {
+                opt.SemanticConventionVersion = SemanticConventionVersion.Latest;
+            });
+    }
+
+    public void Dispose()
+    {
+        // reset Semantic Convention to default
+        Sdk.CreateTracerProviderBuilder()
+            .AddAWSLambdaConfigurations();
+    }
+
     [Fact]
     public void GetHttpTags_APIGatewayProxyRequest_ReturnsCorrectTags()
     {
@@ -40,11 +57,12 @@ public class AWSLambdaHttpUtilsTests
 
         var expectedTags = new Dictionary<string, object>
         {
-            { "http.scheme", "https" },
-            { "http.target", "/path/test?q1=value1" },
-            { "net.host.name", "localhost" },
-            { "net.host.port", 1234 },
-            { "http.method", "GET" },
+            { ExpectedSemanticConventions.AttributeHttpScheme, "https" },
+            { ExpectedSemanticConventions.AttributeUrlPath, "/path/test" },
+            { ExpectedSemanticConventions.AttributeUrlQuery, "?q1=value1" },
+            { ExpectedSemanticConventions.AttributeNetHostName, "localhost" },
+            { ExpectedSemanticConventions.AttributeNetHostPort, 1234 },
+            { ExpectedSemanticConventions.AttributeHttpMethod, "GET" },
         };
 
         AssertTags(expectedTags, actualTags);
@@ -72,11 +90,12 @@ public class AWSLambdaHttpUtilsTests
 
         var expectedTags = new Dictionary<string, object>
         {
-            { "http.scheme", "https" },
-            { "http.target", "/path/test?q1=value1" },
-            { "net.host.name", "localhost" },
-            { "net.host.port", 1234 },
-            { "http.method", "GET" },
+            { ExpectedSemanticConventions.AttributeHttpScheme, "https" },
+            { ExpectedSemanticConventions.AttributeUrlPath, "/path/test" },
+            { ExpectedSemanticConventions.AttributeUrlQuery, "?q1=value1" },
+            { ExpectedSemanticConventions.AttributeNetHostName, "localhost" },
+            { ExpectedSemanticConventions.AttributeNetHostPort, 1234 },
+            { ExpectedSemanticConventions.AttributeHttpMethod, "GET" },
         };
 
         AssertTags(expectedTags, actualTags);
@@ -106,11 +125,12 @@ public class AWSLambdaHttpUtilsTests
 
         var expectedTags = new Dictionary<string, object>
         {
-            { "http.scheme", "https" },
-            { "http.target", "/path/test?q1=value1" },
-            { "net.host.name", "localhost" },
-            { "net.host.port", 1234 },
-            { "http.method", "GET" },
+            { ExpectedSemanticConventions.AttributeHttpScheme, "https" },
+            { ExpectedSemanticConventions.AttributeUrlPath, "/path/test" },
+            { ExpectedSemanticConventions.AttributeUrlQuery, "?q1=value1" },
+            { ExpectedSemanticConventions.AttributeNetHostName, "localhost" },
+            { ExpectedSemanticConventions.AttributeNetHostPort, 1234 },
+            { ExpectedSemanticConventions.AttributeHttpMethod, "GET" },
         };
 
         AssertTags(expectedTags, actualTags);
@@ -132,10 +152,11 @@ public class AWSLambdaHttpUtilsTests
 
         var expectedTags = new Dictionary<string, object>
         {
-            { "http.target", string.Empty },
-            { "http.scheme", "http" },
-            { "net.host.name", "myhost" },
-            { "net.host.port", 432 },
+            { ExpectedSemanticConventions.AttributeUrlPath, string.Empty },
+            { ExpectedSemanticConventions.AttributeUrlQuery, string.Empty },
+            { ExpectedSemanticConventions.AttributeHttpScheme, "http" },
+            { ExpectedSemanticConventions.AttributeNetHostName, "myhost" },
+            { ExpectedSemanticConventions.AttributeNetHostPort, 432 },
         };
 
         AssertTags(expectedTags, actualTags);
@@ -160,7 +181,7 @@ public class AWSLambdaHttpUtilsTests
 
         var expectedTags = new Dictionary<string, object>
         {
-            { "http.status_code", 200 },
+            { ExpectedSemanticConventions.AttributeHttpStatusCode, 200 },
         };
 
         var actualTags = activity?.TagObjects
@@ -237,8 +258,9 @@ public class AWSLambdaHttpUtilsTests
 
         var expectedTags = new Dictionary<string, object>
         {
-            { "http.method", "POST" },
-            { "http.target", "/path/test?q1=value1" },
+            { ExpectedSemanticConventions.AttributeHttpMethod, "POST" },
+            { ExpectedSemanticConventions.AttributeUrlPath, "/path/test" },
+            { ExpectedSemanticConventions.AttributeUrlQuery, "?q1=value1" },
         };
 
         AssertTags(expectedTags, actualTags);
@@ -260,10 +282,11 @@ public class AWSLambdaHttpUtilsTests
 
         var expectedTags = new Dictionary<string, object>
         {
-            { "http.target", string.Empty },
-            { "http.scheme", "http" },
-            { "net.host.name", "myhost" },
-            { "net.host.port", 432 },
+            { ExpectedSemanticConventions.AttributeUrlPath, string.Empty },
+            { ExpectedSemanticConventions.AttributeUrlQuery, string.Empty },
+            { ExpectedSemanticConventions.AttributeHttpScheme, "http" },
+            { ExpectedSemanticConventions.AttributeNetHostName, "myhost" },
+            { ExpectedSemanticConventions.AttributeNetHostPort, 432 },
         };
 
         AssertTags(expectedTags, actualTags);
@@ -294,11 +317,12 @@ public class AWSLambdaHttpUtilsTests
 
         var expectedTags = new Dictionary<string, object>
         {
-            { "http.scheme", "https" },
-            { "http.target", "/path/test?q1=value1" },
-            { "net.host.name", "localhost" },
-            { "net.host.port", 1234 },
-            { "http.method", "GET" },
+            { ExpectedSemanticConventions.AttributeHttpScheme, "https" },
+            { ExpectedSemanticConventions.AttributeUrlPath, "/path/test" },
+            { ExpectedSemanticConventions.AttributeUrlQuery, "?q1=value1" },
+            { ExpectedSemanticConventions.AttributeNetHostName, "localhost" },
+            { ExpectedSemanticConventions.AttributeNetHostPort, 1234 },
+            { ExpectedSemanticConventions.AttributeHttpMethod, "GET" },
         };
 
         AssertTags(expectedTags, actualTags);
@@ -320,10 +344,11 @@ public class AWSLambdaHttpUtilsTests
 
         var expectedTags = new Dictionary<string, object>
         {
-            { "http.target", string.Empty },
-            { "http.scheme", "http" },
-            { "net.host.name", "myhost" },
-            { "net.host.port", 432 },
+            { ExpectedSemanticConventions.AttributeUrlPath, string.Empty },
+            { ExpectedSemanticConventions.AttributeUrlQuery, string.Empty },
+            { ExpectedSemanticConventions.AttributeHttpScheme, "http" },
+            { ExpectedSemanticConventions.AttributeNetHostName, "myhost" },
+            { ExpectedSemanticConventions.AttributeNetHostPort, 432 },
         };
 
         AssertTags(expectedTags, actualTags);
@@ -348,7 +373,7 @@ public class AWSLambdaHttpUtilsTests
 
         var expectedTags = new Dictionary<string, object>
         {
-            { "http.status_code", 200 },
+            { ExpectedSemanticConventions.AttributeHttpStatusCode, 200 },
         };
 
         Assert.NotNull(activity);
@@ -377,7 +402,7 @@ public class AWSLambdaHttpUtilsTests
 
         var expectedTags = new Dictionary<string, object>
         {
-            { "http.status_code", 200 },
+            { ExpectedSemanticConventions.AttributeHttpStatusCode, 200 },
         };
 
         var actualTags = activity?.TagObjects
@@ -453,5 +478,16 @@ public class AWSLambdaHttpUtilsTests
         {
             Assert.Contains(new KeyValuePair<string, TActualValue>(tag.Key, (TActualValue)tag.Value), actualTags);
         }
+    }
+
+    private static class ExpectedSemanticConventions
+    {
+        public const string AttributeHttpScheme = "url.scheme";
+        public const string AttributeUrlPath = "url.path";
+        public const string AttributeUrlQuery = "url.query";
+        public const string AttributeNetHostName = "server.address";
+        public const string AttributeNetHostPort = "server.port";
+        public const string AttributeHttpMethod = "http.request.method";
+        public const string AttributeHttpStatusCode = "http.response.status_code";
     }
 }